Convert float numerics to double

+++ RealTimeWeatherMonitoringApp/Domain/Models/Condition.cs
@@ -4,9 +4,9 @@ public class Condition
 {
     public string Type { get; set; }
     public string Operator { get; set; }
-    public float Value { get; set; }
+    public double Value { get; set; }
 
-    public Condition(string type, string @operator, float value)
+    public Condition(string type, string @operator, double value)
     {
         Type = type;
         Operator = @operator;
